initial project commit
This is the initial project commit.  I still have lots to do both in
adding new features and in restructuring some of my code to be a bit
more efficient.  The to-do list in the main directory should contain
most of the things I still need to do.  Also, at some point I want to do
an in-depth readme file.